pyproject.toml
[project]
name = "pgschema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo",
    "pandas",
    "flask",
    "psycopg2-binary",
    "dotenv>=0.9.9",
    "python-dotenv>=1.1.1",
    "pymysql>=1.1.1",
    "mysql-connector-python>=8.0.28",
]
